+++ scripts/install-mongo.sh
@@ -2,6 +2,15 @@
 
 # MongoDB
 
+if [ -f /home/vagrant/.hercules-installed/mongodb ]
+then
+    echo "MongoDB is already installed"
+    exit 0
+fi
+
+touch /home/vagrant/.hercules-installed/mongodb
+
+
 ## Install
 sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 9DA31620334BD75D9DCB49F368818C72E52529D4
 echo "deb [ arch=amd64,arm64 ] https://repo.mongodb.org/apt/ubuntu xenial/mongodb-org/4.0 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-org-4.0.list